Migrate to Automa v1
This change is not entirely straightforward. One of the major changes in Automa
v1 is the detection of ambiguos NFAs at compile time. Automa's behaviour in the
presence of such NFAs is undefined.
It turns out that the existing BED Machines were ambiguous, which means there
probably was some latent bugs waiting to happen.
To mitigate these, the Machine's behaviour has been slightly changed.
In particular:
* Chrom now cannot start with # or with whitespace, to distinguish them from
  comments and whitespace
* Blank lines now no longer can start with tabs, to distinguish them from a
  record with an empty chrom name

I'm guessing this will not cause any issues in practise.